Fixes #2717 When disabling a Rule, directives or groups can be deleted from it
The form (RuleEditForm) is created with a rule as a val arg. But when we
update the rule by adding a directive for example, and then changing the
status to disable, the val arg rule has not change.

The solution is to use a var rule arg for RuleEditForm, and update it
after a save.