Fixes #4160: update minor version of library dependencies
Normally, all these updates are transparent, safe one: Scala
is stricter in 2.10.3 on overloaded method signature (it was
a bug to not be so), and so, we need to never ever use again
the logback syntax for logger: logger("...{} ... {} ..", foo, bar)

Also, the maven plugin use the "compat" property, allowing to remove
unecessary warnings.

Finally, compilation times should be quicker - and we don't use
yet Zinc and JVM tieredCompilation!