Fixes #2787 Use Doobie in place of Squeryl and most of JDBC
This commit replace all of Squeryl by doobie, and most
of JDBC. It may be commited as it is, as it's a major
improvement, and the remaining bit of jdbc can be
removed after that first big merge.

All the tests pass, some were added to check
more queries.
The squeryl port was rather hard, because there is a lot
of difference between it and doobie.
The change from jdbc is much more easier, because it
basically mean, for a first pass, "keep the jdbc,
remove the mapping which mostly infered, add some types,
remove transformation to Scala collection which is given.
It also allows a much more simple management of transaction,
which are explicits in doobie.

All in all, it makes writting SQL rather pleasant (but for
the compilation times which seems to rocket upward...)