Fixes #20095: If a rule is in a category that no longer exists, it can't be accessed in rule tree
Fixes #20095: If a rule is in a category that no longer exists, it can't be accessed in rule tree

fixup! Fixes #20095: If a rule is in a category that no longer exists, it can't be accessed in rule tree

Fixes #20095: If a rule is in a category that no longer exists, it can't be accessed in rule tree

fixup! fixup! Fixes #20095: If a rule is in a category that no longer exists, it can't be accessed in rule tree

Fixes #20095: If a rule is in a category that no longer exists, it can't be accessed in rule tree